requirements.txt
masumi
python-dotenv
openai
pytrends
fastapi
asyncpg
